whatsapp_bot/requirements.txt
flask
requests
gspread
oauth2client
razorpay
python-dotenv
